+++ R/amlrPinnipeds-package.R
@@ -34,7 +34,7 @@
 #' @importFrom stringi stri_escape_unicode
 #' @importFrom stringr str_length str_match str_remove_all str_replace_all
 #' str_replace_na str_sub str_to_lower str_to_sentence
-#' @importFrom tidyr complete nest unnest pivot_longer pivot_wider replace_na
+#' @importFrom tidyr complete nest nesting unnest pivot_longer pivot_wider replace_na
 #' @importFrom utils globalVariables read.csv write.csv
 #'
 #' @keywords package